test code for agents to figure out what tile they reside in

+++ src/SwarmManager.h
@@ -57,6 +57,6 @@ private:
 
     int grid_size_x;
     int grid_size_y;
-    float grid_scale;
+    int grid_scale;
 
 };
